+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 0.2.0:
   * Make `Debug` string representations more compact.
+  * `Month::wrapping_add/sub` now take an `i8` instead of `u8`.
 
 0.1.1:
   * Fix `Date::add_days` for days in years 100, 200 and 300 of the 400 year cycle.